LayoutTests/media/media-can-play-mpeg-audio.html

<!DOCTYPE html>
<html>
    <head>
        <script src=video-test.js></script>
        <script>
            function start()
            {
                audio = document.getElementsByTagName('audio')[0];

                testExpected("audio.canPlayType('audio/x-mpeg')", "maybe");
                testExpected("audio.canPlayType('audio/x-mp3')", "maybe");
                testExpected("audio.canPlayType('audio/mp3')", "maybe");
                testExpected("audio.canPlayType('audio/mpeg')", "maybe");

                testExpected("audio.canPlayType('audio/mp4')", "maybe");
                testExpected("audio.canPlayType('audio/aac')", "maybe");

                endTest();
            }
        </script>
    </head>
    <body onload="start()">
        <audio controls></audio>
        <p>Test HTMLMediaElement <em>canPlayType()</em> method with multiple audio mpeg MIME types.</p>
        <p>These tests may be expected to fail if the WebKit port does not support the format.</p>
    </body>
</html>
